feat: refactor and updates
