+++ apps/watch/src/components/Videos/Videos.stories.tsx
@@ -3,7 +3,6 @@ import { MockedProvider } from '@apollo/client/testing'
 import { watchConfig } from '../../libs/storybook'
 import { videos } from './testData'
 import { GET_VIDEOS, Videos } from './Videos'
-import { CarouselItem } from './VideosCarousel/CarouselItem'
 
 const VideosStory = {
   ...watchConfig,
@@ -43,29 +42,6 @@ const Template: Story = ({ ...args }) => {
   )
 }
 
-const Tester: Story = ({ ...args }) => {
-  return (
-    <CarouselItem
-      title={args.title}
-      image={args.image}
-      variant={args.variant}
-      videoType="standalone"
-      isPlaying={false}
-      onClick={() => {
-        console.log('clicked it')
-      }}
-    />
-  )
-}
-
-export const CarouselItemTest = Tester.bind({})
-CarouselItemTest.args = {
-  title: [{ value: 'Testing' }],
-  image:
-    'https://d1wl257kev7hsz.cloudfront.net/cinematics/1_cl-0-0.mobileCinematicHigh.jpg',
-  variant: { duration: 3680, __typename: 'VideoVariant' }
-}
-
 export const Carousel = Template.bind({})
 Carousel.args = {
   layout: 'carousel',
@@ -77,5 +53,4 @@ Grid.args = {
   layout: 'grid',
   limit: 20
 }
-
 export default VideosStory as Meta
